Add server client and send/receive functionality
Added a client for the BingoServer. The BoardWindow sends and receives clicks through abstracted BingoReceiver and BingoSender classes. These classes interface directly with the BingoServer and BingoClient classes so the BoardWindow does not need to know whether it is the server or client. All it needs to do is call sender->Send() when it needs to send data. Whenever the BingoReceiver receives data, it emits a Received signal which the BoardWindow is connected to.